% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landtrendr.R
\name{buildMosaic}
\alias{buildMosaic}
\title{Bluid Mosaic Function}
\usage{
buildMosaic(year, startDay, endDay, aoi, dummyCollection)
}
\arguments{
\item{year}{this parameter is automatically and refers to each year from Start Year and End Year}

\item{startDay}{String variable - date tie to the Starting year (Default: '06-01')}

\item{endDay}{String variable - date tie to the Ending year (Default: '09-30')}

\item{aoi}{this parameter refers Area Of Interest and constrain the retrieve images to the AOI.}

\item{dummyCollection}{empty ee.ImageCollection filled with zeros}
}
\value{
ee.Image
}
\description{
This function creates a mosaic from ee.Image
}
