% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landtrendr.R
\name{buildMosaicCollection}
\alias{buildMosaicCollection}
\title{Bluid Mosaic Collection Function}
\usage{
buildMosaicCollection(
  startYear,
  endYear,
  startDay,
  endDay,
  aoi,
  dummyCollection
)
}
\arguments{
\item{startYear}{Numeric variable - Starting year of th analysis (Default: 2010)}

\item{endYear}{Numeric variable - Ending year of th analysis (Default: 2017)}

\item{startDay}{String variable - date tie to the Starting year (Default: '06-01')}

\item{endDay}{String variable - date tie to the Ending year (Default: '09-30')}

\item{aoi}{this parameter refers Area Of Interest and constrain the retrieve images to the AOI.}

\item{dummyCollection}{empty ee.ImageCollection filled with zeros}
}
\value{
ee.Image
}
\description{
This function creates an Image collection with 6 bands
}
