% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landtrendr.R
\name{calcDifFromMed}
\alias{calcDifFromMed}
\title{Difference calculation from Medoid function}
\usage{
calcDifFromMed(img)
}
\arguments{
\item{img}{ee.Image}
}
\value{
reduced Image ee_Image
}
\description{
This function change dynamically and calculates de Difference image from Medoid fuction
}
