% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landtrendr.R
\name{getCombinedSRcollection}
\alias{getCombinedSRcollection}
\title{SR Collection merge}
\usage{
getCombinedSRcollection(year, startDay, endDay, aoi)
}
\arguments{
\item{year}{this parameter is automatically and refers to each year from Start Year and End Year}

\item{startDay}{this parameter refers to month and day of the starting year}

\item{endDay}{this parameter refers to month and day of the ending year}

\item{aoi}{this parameter refers Area Of Interest and constrain the retrieve images to the AOI.}
}
\description{
This function merge the outcome Image collection from each Landsat mission.
}
