% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landtrendr.R
\name{getLTvertStack}
\alias{getLTvertStack}
\title{LTvertStack Function}
\usage{
getLTvertStack(LTresult)
}
\arguments{
\item{LTresult}{LandtrendR result}
}
\value{
ee.Image
}
\description{
This function extracts the segmentation vertex info
}
