% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landtrendr.R
\name{global_variable}
\alias{global_variable}
\alias{global_variables}
\title{Global variables setup}
\usage{
global_variable()
}
\description{
In this section the global parameters for initial config are setup
}
