% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landtrendr.R
\name{harmonizationRoy}
\alias{harmonizationRoy}
\title{Harmonize L8 images}
\usage{
harmonizationRoy(oli)
}
\arguments{
\item{oli}{L8 Image in ee.Image format.}
}
\value{
oli_harmonized ee.Image object
}
\description{
This function harmonized the L8 image as the sensor has different specification as L5 and L7
}
