% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landtrendr.R
\name{invertIndex}
\alias{invertIndex}
\title{Invert index Function}
\usage{
invertIndex(img)
}
\arguments{
\item{img}{ee.Image}
}
\value{
ee.Image
}
\description{
This function invert the direction of the magnitud to make it logic for the user.
}
