% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landtrendr.R
\name{LtRun}
\alias{LtRun}
\title{Run LandTrendr Algorithm}
\usage{
LtRun(
  startYear = 2010,
  endYear = 2017,
  startDay = "06-01",
  endDay = "09-30",
  aoi = aoi_,
  runParams = runParam
)
}
\arguments{
\item{startYear}{Numeric variable - Starting year of th analysis (Default: 2010)}

\item{endYear}{Numeric variable - Ending year of th analysis (Default: 2017)}

\item{startDay}{String variable - date tie to the Starting year (Default: '06-01')}

\item{endDay}{String variable - date tie to the Ending year (Default: '09-30')}

\item{aoi}{Polygon sf Geometry - Area of Interest (Default: US County from sf dataset)}

\item{runParams}{Parameters for Landtrendr Analysis - By default the  package take the following values: maxSegments = 6, spikeThreshold = 0.9, vertexCountOvershoot = 3, preventOneYearRecovery =  TRUE,recoveryThreshold = 0.25, pvalThreshold = 0.05, bestModelProportion = 0.75, minObservationsNeeded= 6}
}
\value{
lt ee.Image object
}
\description{
This function built a ImageCollection suitable for LT analysis, using Harmonization and pre-processes before running the Landtrendr Algorithms
The function starts building the image and an additional NBR index that helps the algorithm to calculate the magnitude of change.
This function is based on rgee package, thus the outcome of this function is a ee.Image object.
}
