% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landtrendr.R
\name{medoidMosaic}
\alias{medoidMosaic}
\title{Medoid Function}
\usage{
medoidMosaic(inCollection, dummyCollection)
}
\arguments{
\item{inCollection}{ee.ImageCollection from the range of dates provided}

\item{dummyCollection}{empty ee.ImageCollection filled with zeros}
}
\value{
Image Collection
}
\description{
This function creates an Image collection with 6 bands
}
