% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landtrendr.R
\docType{data}
\name{pkg.globals}
\alias{pkg.globals}
\alias{initial_variables}
\title{Initial test}
\format{
An object of class \code{environment} of length 10.
}
\usage{
pkg.globals
}
\arguments{
\item{aoi_}{Initial Area Of Interest}

\item{distDir}{Parameter of Direction}

\item{global_sensor}{parameter of the sensor to use (e.g L08)}

\item{global_median_calcDif}{mean calculated based on the Collection}

\item{indexNameFTV}{Initialize the FTV param}

\item{runParam}{Initial config for segmentation process}

\item{dummyCollection}{Initialize the a collection of zeros}
}
\description{
In this section the global parameters for initial setup
}
\keyword{datasets}
