% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landtrendr.R
\name{prepImages}
\alias{prepImages}
\title{Image preparation function}
\usage{
prepImages(img)
}
\arguments{
\item{img}{ee.Image variable that refers to images in the preselected image collection based on input parameter od LtRun}
}
\description{
This function mask the images in the Image collection built with the LtRun input varibles.
}
