% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landtrendr.R
\name{segIndex}
\alias{segIndex}
\title{segIndex Function}
\usage{
segIndex(img)
}
\arguments{
\item{img}{ee.Image}
}
\value{
ee.Image
}
\description{
This function calculates and add to the Image the calculation of NBR index
}
