% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landtrendr.R
\name{set_global_median_calcDif}
\alias{set_global_median_calcDif}
\title{Global Median Difference}
\usage{
set_global_median_calcDif(finalCollection)
}
\arguments{
\item{finalCollection}{this parameter is automatically and refers to each year image in the construction of mosaics}
}
\description{
This function change dynamically the global median different variable.
}
