% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landtrendr.R
\name{set_global_sensor}
\alias{set_global_sensor}
\title{Dynamic Sensor variable}
\usage{
set_global_sensor(sensor)
}
\arguments{
\item{sensor}{string variable that refers to the sensor name in GEE syntaxis (i.e L08,L07,L05)}
}
\description{
This function change dynamically all the sensor variables during the process.
}
