% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landtrendr.R
\name{setAOI}
\alias{setAOI}
\title{Set a new AOI}
\usage{
setAOI(aoi_)
}
\arguments{
\item{aoi_}{new Area of Interest, should be a sf geometry}
}
\description{
This function allows to set up a new Area of Interest
}
