% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landtrendr.R
\name{setParameters}
\alias{setParameters}
\title{LandTrendR running Parameter}
\usage{
setParameters(param)
}
\arguments{
\item{param}{List of parameter}
}
\description{
The function will change the initial parameters to process the algorithm
}
\examples{
list( maxSegments = 6,
spikeThreshold = 0.9,
vertexCountOvershoot = 3,
preventOneYearRecovery =  TRUE,
recoveryThreshold = 0.25,
pvalThreshold = 0.05,
bestModelProportion = 0.75,
minObservationsNeeded= 6)
}
